created libnavigation-thirdparty module
