+++ libnavigation-core/build.gradle
@@ -60,6 +60,7 @@ dependencies {
     testImplementation dependenciesList.androidXCore
 
     testImplementation project(':libtesting-utils')
+    testImplementation project(':libtesting-thirdparty')
     apply from: "${rootDir}/gradle/unit-testing-dependencies.gradle"
     testImplementation dependenciesList.commonsIO
