+++ libtesting-thirdparty/src/main/java/com/mapbox/navigation/testing/factories/DirectionsResponseFactories.kt
@@ -1,4 +1,4 @@
-package com.mapbox.navigation.core.infra.factories
+package com.mapbox.navigation.testing.factories
 
 import com.mapbox.api.directions.v5.DirectionsCriteria
 import com.mapbox.api.directions.v5.models.DirectionsRoute
@@ -7,42 +7,7 @@ import com.mapbox.api.directions.v5.models.LegAnnotation
 import com.mapbox.api.directions.v5.models.MaxSpeed
 import com.mapbox.api.directions.v5.models.RouteLeg
 import com.mapbox.api.directions.v5.models.RouteOptions
-import com.mapbox.bindgen.Expected
-import com.mapbox.bindgen.ExpectedFactory
 import com.mapbox.geojson.Point
-import com.mapbox.navigation.base.internal.SDKRouteParser
-import com.mapbox.navigation.base.internal.utils.mapToNativeRouteOrigin
-import com.mapbox.navigation.base.route.NavigationRoute
-import com.mapbox.navigation.base.route.RouterOrigin
-import com.mapbox.navigator.RouteInterface
-
-fun createNavigationRoute(
-    directionsRoute: DirectionsRoute = createDirectionsRoute()
-): NavigationRoute {
-    return com.mapbox.navigation.base.internal.route.createNavigationRoute(
-        directionsRoute,
-        object : SDKRouteParser {
-            override fun parseDirectionsResponse(
-                response: String,
-                request: String,
-                routerOrigin: RouterOrigin
-            ): Expected<String, List<RouteInterface>> {
-                return ExpectedFactory.createValue(
-                    listOf(
-                        createRouteInterface(
-                            responseUUID = directionsRoute.requestUuid() ?: "null",
-                            routeIndex = directionsRoute.routeIndex()!!.toInt(),
-                            responseJson = response,
-                            routerOrigin = routerOrigin.mapToNativeRouteOrigin(),
-                            requestURI = directionsRoute.routeOptions()!!.toUrl("test")
-                                .toString()
-                        ),
-                    )
-                )
-            }
-        }
-    )
-}
 
 fun createDirectionsRoute(
     legs: List<RouteLeg>? = listOf(createRouteLeg()),
