+++ libtesting-thirdparty/src/main/java/com/mapbox/navigation/testing/factories/NativeFactories.kt
@@ -1,4 +1,4 @@
-package com.mapbox.navigation.core.infra.factories
+package com.mapbox.navigation.testing.factories
 
 import com.mapbox.geojson.Point
 import com.mapbox.navigator.ActiveGuidanceInfo
