+++ settings.gradle
@@ -25,3 +25,4 @@ include ':examples',
         ':libnavui-androidauto',
         ':libnavui-app',
         ':qa-test-app'
+include ':libtesting-thirdparty'
